+++ tests/functional/functional_test_case_trait.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * Simple Spoiler extension for phpBB.
+ * @author Alfredo Ramos <alfredo.ramos@yandex.com>
+ * @copyright 2017 Alfredo Ramos
+ * @license GPL-2.0-only
+ */
+
+namespace alfredoramos\simplespoiler\tests\functional;
+
+trait functional_test_case_trait
+{
+	static protected function setup_extensions()
+	{
+		return ['alfredoramos/simplespoiler'];
+	}
+}
